move showWalkthroughModal into own module
