+++ src/components/quickBar/defaultActions.tsx
@@ -29,7 +29,7 @@ import { faSlack } from "@fortawesome/free-brands-svg-icons";
 import { DEFAULT_SERVICE_URL, MARKETPLACE_URL } from "@/urlConstants";
 import { Events } from "@/telemetry/events";
 import reportEvent from "@/telemetry/reportEvent";
-import { showWalkthroughModal } from "@/components/walkthroughModal/WalkthroughModalApp";
+import showWalkthroughModal from "@/components/walkthroughModal/showWalkthroughModal";
 
 const PIXIEBRIX_SECTION = "PixieBrix";
